Refactor DataRequest class to use new RequestV2 model
Replaced the usage of RequestInfo with the new RequestV2 model in the DataRequest class. This includes changes to the way instances of these classes are created, queried and transformed to dictionary. This refactoring aligns with the introduction of the new RequestV2 model and makes the DataRequest class compatible with it.